Refactor: 멘토님 리뷰 반영 - 컴포넌트 다른 파일로 분리
