+++ src/Pages/Checkout/Checkout.scss
@@ -189,20 +189,6 @@
           }
         }
       }
-      .selectCard {
-      }
-      .installmentPeriod {
-        display: flex;
-        align-items: center;
-        .selectInstallment {
-        }
-      }
-      .transferToWhom {
-        input {
-          width: 100%;
-          max-width: 300px;
-        }
-      }
     }
   }
   .orderSummary {
