+++ src/Pages/Checkout/Virtual.scss
@@ -1,0 +1,2 @@
+@import "../../Styles/common.scss";
+@use 'Checkout';
